🐛 FIX: Huitalcoche now can be played on slot with max cards
